+++ libraries/common/src/main/java/androidx/media3/common/Player.java
@@ -242,7 +242,7 @@ public interface Player {
   }
 
   /** Position info describing a playback position involved in a discontinuity. */
-  final class PositionInfo implements Bundleable {
+  final class PositionInfo {
 
     /**
      * The UID of the window, or {@code null} if the timeline is {@link Timeline#isEmpty() empty}.
@@ -385,8 +385,6 @@ public interface Player {
           && Objects.equal(mediaItem, other.mediaItem);
     }
 
-    // Bundleable implementation.
-
     @VisibleForTesting static final String FIELD_MEDIA_ITEM_INDEX = Util.intToStringMaxRadix(0);
     private static final String FIELD_MEDIA_ITEM = Util.intToStringMaxRadix(1);
     @VisibleForTesting static final String FIELD_PERIOD_INDEX = Util.intToStringMaxRadix(2);
@@ -426,10 +424,11 @@ public interface Player {
     }
 
     /**
-     * {@inheritDoc}
+     * Returns a {@link Bundle} representing the information stored in this object.
      *
      * <p>It omits the {@link #windowUid} and {@link #periodUid} fields. The {@link #windowUid} and
-     * {@link #periodUid} of an instance restored by {@link #CREATOR} will always be {@code null}.
+     * {@link #periodUid} of an instance restored by {@link #fromBundle(Bundle)} will always be
+     * {@code null}.
      *
      * @param controllerInterfaceVersion The interface version of the media controller this Bundle
      *     will be sent to.
@@ -462,26 +461,16 @@ public interface Player {
     }
 
     /**
-     * {@inheritDoc}
+     * Returns a {@link Bundle} representing the information stored in this object.
      *
      * <p>It omits the {@link #windowUid} and {@link #periodUid} fields. The {@link #windowUid} and
-     * {@link #periodUid} of an instance restored by {@link #CREATOR} will always be {@code null}.
+     * {@link #periodUid} of an instance restored by {@link #fromBundle} will always be {@code
+     * null}.
      */
     @UnstableApi
-    @Override
     public Bundle toBundle() {
       return toBundle(Integer.MAX_VALUE);
     }
-
-    /**
-     * Object that can restore {@link PositionInfo} from a {@link Bundle}.
-     *
-     * @deprecated Use {@link #fromBundle} instead.
-     */
-    @UnstableApi
-    @Deprecated
-    @SuppressWarnings("deprecation") // Deprecated instance of deprecated class
-    public static final Creator<PositionInfo> CREATOR = PositionInfo::fromBundle;
 
     /** Restores a {@code PositionInfo} from a {@link Bundle}. */
     @UnstableApi
@@ -514,7 +503,7 @@ public interface Player {
    *
    * <p>Instances are immutable.
    */
-  final class Commands implements Bundleable {
+  final class Commands {
 
     /** A builder for {@link Commands} instances. */
     @UnstableApi
@@ -744,12 +733,9 @@ public interface Player {
       return flags.hashCode();
     }
 
-    // Bundleable implementation.
-
     private static final String FIELD_COMMANDS = Util.intToStringMaxRadix(0);
 
     @UnstableApi
-    @Override
     public Bundle toBundle() {
       Bundle bundle = new Bundle();
       ArrayList<Integer> commandsBundle = new ArrayList<>();
@@ -759,16 +745,6 @@ public interface Player {
       bundle.putIntegerArrayList(FIELD_COMMANDS, commandsBundle);
       return bundle;
     }
-
-    /**
-     * Object that can restore {@link Commands} from a {@link Bundle}.
-     *
-     * @deprecated Use {@link #fromBundle} instead.
-     */
-    @UnstableApi
-    @Deprecated
-    @SuppressWarnings("deprecation") // Deprecated instance of deprecated class
-    public static final Creator<Commands> CREATOR = Commands::fromBundle;
 
     /** Restores a {@code Commands} from a {@link Bundle}. */
     @UnstableApi