+++ libraries/common/src/main/java/androidx/media3/common/Timeline.java
@@ -139,7 +139,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
  * <p>This case includes mid-roll ad groups, which are defined as part of the timeline's single
  * period. The period can be queried for information about the ad groups and the ads they contain.
  */
-public abstract class Timeline implements Bundleable {
+public abstract class Timeline {
 
   /**
    * Holds information about a window in a {@link Timeline}. A window usually corresponds to one
@@ -152,7 +152,7 @@ public abstract class Timeline implements Bundleable {
    * src="https://developer.android.com/static/images/reference/androidx/media3/common/timeline-window.svg"
    * alt="Information defined by a timeline window">
    */
-  public static final class Window implements Bundleable {
+  public static final class Window {
 
     /**
      * A {@link #uid} for a window that must be used for single-window {@link Timeline Timelines}.
@@ -412,8 +412,6 @@ public abstract class Timeline implements Bundleable {
       return result;
     }
 
-    // Bundleable implementation.
-
     private static final String FIELD_MEDIA_ITEM = Util.intToStringMaxRadix(1);
     private static final String FIELD_PRESENTATION_START_TIME_MS = Util.intToStringMaxRadix(2);
     private static final String FIELD_WINDOW_START_TIME_MS = Util.intToStringMaxRadix(3);
@@ -430,14 +428,13 @@ public abstract class Timeline implements Bundleable {
     private static final String FIELD_POSITION_IN_FIRST_PERIOD_US = Util.intToStringMaxRadix(13);
 
     /**
-     * {@inheritDoc}
+     * Returns a {@link Bundle} representing the information stored in this object.
      *
      * <p>It omits the {@link #uid} and {@link #manifest} fields. The {@link #uid} of an instance
-     * restored by {@link #CREATOR} will be a fake {@link Object} and the {@link #manifest} of the
-     * instance will be {@code null}.
+     * restored by {@link #fromBundle} will be a fake {@link Object} and the {@link #manifest} of
+     * the instance will be {@code null}.
      */
     @UnstableApi
-    @Override
     public Bundle toBundle() {
       Bundle bundle = new Bundle();
       if (!MediaItem.EMPTY.equals(mediaItem)) {
@@ -483,19 +480,6 @@ public abstract class Timeline implements Bundleable {
       }
       return bundle;
     }
-
-    /**
-     * Object that can restore {@link Period} from a {@link Bundle}.
-     *
-     * <p>The {@link #uid} of a restored instance will be a fake {@link Object} and the {@link
-     * #manifest} of the instance will be {@code null}.
-     *
-     * @deprecated Use {@link #fromBundle} instead.
-     */
-    @UnstableApi
-    @Deprecated
-    @SuppressWarnings("deprecation") // Deprecated instance of deprecated class
-    public static final Creator<Window> CREATOR = Window::fromBundle;
 
     /** Restores a {@code Window} from a {@link Bundle}. */
     @UnstableApi
@@ -559,7 +543,7 @@ public abstract class Timeline implements Bundleable {
    * src="https://developer.android.com/static/images/reference/androidx/media3/common/timeline-period.svg"
    * alt="Information defined by a period">
    */
-  public static final class Period implements Bundleable {
+  public static final class Period {
 
     /**
      * An identifier for the period. Not necessarily unique. May be null if the ids of the period
@@ -908,8 +892,6 @@ public abstract class Timeline implements Bundleable {
       return result;
     }
 
-    // Bundleable implementation.
-
     private static final String FIELD_WINDOW_INDEX = Util.intToStringMaxRadix(0);
     private static final String FIELD_DURATION_US = Util.intToStringMaxRadix(1);
     private static final String FIELD_POSITION_IN_WINDOW_US = Util.intToStringMaxRadix(2);
@@ -917,13 +899,12 @@ public abstract class Timeline implements Bundleable {
     private static final String FIELD_AD_PLAYBACK_STATE = Util.intToStringMaxRadix(4);
 
     /**
-     * {@inheritDoc}
+     * Returns a {@link Bundle} representing the information stored in this object.
      *
      * <p>It omits the {@link #id} and {@link #uid} fields so these fields of an instance restored
-     * by {@link #CREATOR} will always be {@code null}.
+     * by {@link #fromBundle} will always be {@code null}.
      */
     @UnstableApi
-    @Override
     public Bundle toBundle() {
       Bundle bundle = new Bundle();
       if (windowIndex != 0) {
@@ -943,18 +924,6 @@ public abstract class Timeline implements Bundleable {
       }
       return bundle;
     }
-
-    /**
-     * Object that can restore {@link Period} from a {@link Bundle}.
-     *
-     * <p>The {@link #id} and {@link #uid} of restored instances will always be {@code null}.
-     *
-     * @deprecated Use {@link #fromBundle} instead.
-     */
-    @UnstableApi
-    @Deprecated
-    @SuppressWarnings("deprecation") // Deprecated instance of deprecated class
-    public static final Creator<Period> CREATOR = Period::fromBundle;
 
     /** Restores a {@code Period} from a {@link Bundle}. */
     @UnstableApi
@@ -1398,21 +1367,18 @@ public abstract class Timeline implements Bundleable {
     return result;
   }
 
-  // Bundleable implementation.
-
   private static final String FIELD_WINDOWS = Util.intToStringMaxRadix(0);
   private static final String FIELD_PERIODS = Util.intToStringMaxRadix(1);
   private static final String FIELD_SHUFFLED_WINDOW_INDICES = Util.intToStringMaxRadix(2);
 
   /**
-   * {@inheritDoc}
+   * Returns a {@link Bundle} representing the information stored in this object.
    *
    * <p>The {@link #getWindow(int, Window)} windows} and {@link #getPeriod(int, Period) periods} of
-   * an instance restored by {@link #CREATOR} may have missing fields as described in {@link
+   * an instance restored by {@link #fromBundle} may have missing fields as described in {@link
    * Window#toBundle()} and {@link Period#toBundle()}.
    */
   @UnstableApi
-  @Override
   public final Bundle toBundle() {
     List<Bundle> windowBundles = new ArrayList<>();
     int windowCount = getWindowCount();
@@ -1470,20 +1436,6 @@ public abstract class Timeline implements Bundleable {
     return new RemotableTimeline(
         ImmutableList.of(window), periods.build(), /* shuffledWindowIndices= */ new int[] {0});
   }
-
-  /**
-   * Object that can restore a {@link Timeline} from a {@link Bundle}.
-   *
-   * <p>The {@link #getWindow(int, Window)} windows} and {@link #getPeriod(int, Period) periods} of
-   * a restored instance may have missing fields as described in {@link Window#CREATOR} and {@link
-   * Period#CREATOR}.
-   *
-   * @deprecated Use {@link #fromBundle} instead.
-   */
-  @UnstableApi
-  @Deprecated
-  @SuppressWarnings("deprecation") // Deprecated instance of deprecated class
-  public static final Creator<Timeline> CREATOR = Timeline::fromBundle;
 
   /** Restores a {@code Timeline} from a {@link Bundle}. */
   @UnstableApi