+++ libraries/common/src/main/java/androidx/media3/common/TrackGroup.java
@@ -48,7 +48,7 @@ import java.util.List;
  * {@link Tracks.Group}, it does not include runtime information such as the extent to which
  * playback of each track is supported by the device, or which tracks are currently selected.
  */
-public final class TrackGroup implements Bundleable {
+public final class TrackGroup {
 
   private static final String TAG = "TrackGroup";
 
@@ -161,12 +161,10 @@ public final class TrackGroup implements Bundleable {
     return id.equals(other.id) && Arrays.equals(formats, other.formats);
   }
 
-  // Bundleable implementation.
   private static final String FIELD_FORMATS = Util.intToStringMaxRadix(0);
   private static final String FIELD_ID = Util.intToStringMaxRadix(1);
 
   @UnstableApi
-  @Override
   public Bundle toBundle() {
     Bundle bundle = new Bundle();
     ArrayList<Bundle> arrayList = new ArrayList<>(formats.length);
@@ -177,16 +175,6 @@ public final class TrackGroup implements Bundleable {
     bundle.putString(FIELD_ID, id);
     return bundle;
   }
-
-  /**
-   * Object that can restore {@code TrackGroup} from a {@link Bundle}.
-   *
-   * @deprecated Use {@link #fromBundle} instead.
-   */
-  @UnstableApi
-  @Deprecated
-  @SuppressWarnings("deprecation") // Deprecated instance of deprecated class
-  public static final Creator<TrackGroup> CREATOR = TrackGroup::fromBundle;
 
   /** Restores a {@code TrackGroup} from a {@link Bundle}. */
   @UnstableApi